// packages/diagrams/frontend/sirius-components-diagrams/src/renderer/actions/visibility/ManageVisibilityContextProvider.types.ts
/*******************************************************************************
 * Copyright (c) 2025 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/

import { XYPosition } from '@xyflow/react';

export interface ManageVisibilityContextValue {
  openDialog: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>, diagramElementId: string) => void;
  closeDialog: () => void;
}

export interface ManageVisibilityContextState {
  isOpen: boolean;
  diagramElementId: string;
  initialPosition: XYPosition;
}